requirements.txt
# A-LEMS Core Dependencies
streamlit==1.28.0
pandas==2.3.3
plotly==5.17.0
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0
psutil==5.9.8
python-dateutil==2.9.0.post0
tqdm==4.66.5
scipy==1.17.0

# ── DISTRIBUTED (multi-host orchestration) ────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
tomli>=2.0.1

# ── SYSTEM PACKAGES (apt, Oracle VM only) ─────────────────────────────────────
# sudo apt install postgresql postgresql-contrib libpq-dev graphviz
# sudo apt install iptables-persistent netfilter-persistent
